Use references instead of modules
Changed round to be constructed with a pair of Firebase references. This
means that we don't have stale data when we observe modules and changes
are made by other users.

Matchmaker will connect to Firebase and return two references instead of
modules.

Changed EloRating so that we're working with primitve types to
calculate expected win probability and new ratings, instead of module
structs. This makes it easier to work with references.